change type of in_stock col migration

+++ database/migrations/2022_09_26_062935_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('description')->nullable();
             $table->decimal('price', 12, 4);
-            $table->decimal('in_stock',  3, 2);
+            $table->integer('in_stock');
             $table->boolean('active_for_sale', true);
 
             $table->timestamps();
